split agent from watcher and move proxy to a diff file
