+++ pilot/pkg/proxy/agent.go
@@ -58,11 +58,11 @@ import (
 // attempt timers. The call to schedule a configuration update will block until
 // the control loop is ready to accept and process the configuration update.
 type Agent interface {
-	// ScheduleConfigUpdate sets the desired configuration for the proxy.  Agent
-	// compares the current active configuration to the desired state and
+	// ConfigCh returns the config channel used to send configuration updates.
+	// Agent compares the current active configuration to the desired state and
 	// initiates a restart if necessary. If the restart fails, the agent attempts
 	// to retry with an exponential back-off.
-	ScheduleConfigUpdate(config interface{})
+	ConfigCh() chan<- interface{}
 
 	// Run starts the agent control loop and awaits for a signal on the input
 	// channel to exit the loop.
@@ -80,9 +80,9 @@ var (
 )
 
 const (
-	// MaxAborts is the maximum number of cascading abort messages to buffer.
+	// maxAborts is the maximum number of cascading abort messages to buffer.
 	// This should be the upper bound on the number of proxies available at any point in time.
-	MaxAborts = 10
+	maxAborts = 10
 )
 
 // NewAgent creates a new proxy agent for the proxy start-up and clean-up functions.
@@ -157,8 +157,8 @@ type exitStatus struct {
 	err   error
 }
 
-func (a *agent) ScheduleConfigUpdate(config interface{}) {
-	a.configCh <- config
+func (a *agent) ConfigCh() chan<- interface{} {
+	return a.configCh
 }
 
 func (a *agent) Run(ctx context.Context) {
@@ -264,7 +264,7 @@ func (a *agent) reconcile() {
 	// cancel any scheduled restart
 	a.retry.restart = nil
 
-	log.Infof("Reconciling configuration (budget %d)", a.retry.budget)
+	log.Infof("Reconciling retry (budget %d)", a.retry.budget)
 
 	// check that the config is current
 	if reflect.DeepEqual(a.desiredConfig, a.currentConfig) {
@@ -275,15 +275,15 @@ func (a *agent) reconcile() {
 	// discover and increment the latest running epoch
 	epoch := a.latestEpoch() + 1
 	// buffer aborts to prevent blocking on failing proxy
-	abortCh := make(chan error, MaxAborts)
+	abortCh := make(chan error, maxAborts)
 	a.epochs[epoch] = a.desiredConfig
 	a.abortCh[epoch] = abortCh
 	a.currentConfig = a.desiredConfig
-	go a.waitForExit(a.desiredConfig, epoch, abortCh)
+	go a.runWait(a.desiredConfig, epoch, abortCh)
 }
 
-// waitForExit runs the start-up command as a go routine and waits for it to finish
-func (a *agent) waitForExit(config interface{}, epoch int, abortCh <-chan error) {
+// runWait runs the start-up command as a go routine and waits for it to finish
+func (a *agent) runWait(config interface{}, epoch int, abortCh <-chan error) {
 	log.Infof("Epoch %d starting", epoch)
 	err := a.proxy.Run(config, epoch, abortCh)
 	a.statusCh <- exitStatus{epoch: epoch, err: err}
